cephadm: move podman/docker test mock
to use as a default value in the fixtures

(cherry picked from commit 3060962926bc69f03e11352f637d3586143d4ac6)

Conflicts:
	src/cephadm/tests/test_cephadm.py
